Modify how we keep track of dynamic vaults
Previously we had a dynamic vault id. This was troublesome because it would allow a user to have multiple plans that could conflict between them.
Now, dynamic vaults are mapped to the user address. This makes it easier to keep track of the vaults and easier for access control.
Although a vault does not fully allow yet to have multiple owners, this could be possible with some changes.
Other changes:
* Fix the error where the user would not be able to create a testament if he first added a backup address
* Deployed to Mumbai
* Update tests